vendor extension trait
